ai-brain-python/requirements.txt
asyncio
aiokafka
langchain-aws
langchain-core
langgraph
langgraph-checkpoint-redis
langchain
boto3
redis
asyncpg
python-dotenv
httpx